+++ lib/src/handlers/invokers.ts
@@ -1,6 +1,7 @@
 import { invoke as tauriInvoke, InvokeArgs, InvokeOptions } from '@tauri-apps/api/core';
 
 import { Plugin } from '../state/plugin';
+import { Profile } from '../state/profile';
 
 export enum SeelenCommand {
   // General
@@ -35,6 +36,7 @@ export enum SeelenCommand {
   StateGetHistory = 'state_get_history',
   StateGetPlugins = 'state_get_plugins',
   StateGetWidgets = 'state_get_widgets',
+  stateGetProfiles = 'state_get_profiles',
 
   // Media
   MediaPrev = 'media_prev',
@@ -88,6 +90,7 @@ type ReturnTypeByCommand = Record<SeelenCommand, unknown> & {
   [SeelenCommand.CheckForUpdates]: boolean;
   [SeelenCommand.InstallLastAvailableUpdate]: never;
   [SeelenCommand.StateGetPlugins]: Plugin[];
+  [SeelenCommand.stateGetProfiles]: Profile[];
 };
 
 export type SeelenCommandReturn<T extends SeelenCommand> = ReturnTypeByCommand[T];