+++ lib/src/lib.rs
@@ -1,4 +1,7 @@
+pub mod error;
 pub mod rect;
 pub mod state;
 pub mod system_state;
 pub mod handlers;
+
+pub use error::SeelenLibError;