+++ lib/src/state/plugin.rs
@@ -5,4 +5,6 @@ pub struct Plugin {
     pub id: String,
     pub target: String,
     pub plugin: serde_yaml::Value,
+    #[serde(default)]
+    pub bundled: bool,
 }