+++ src/background/utils/mod.rs
@@ -9,7 +9,8 @@ pub use winver::*;
 
 use std::{
     collections::HashMap,
-    path::PathBuf,
+    fs,
+    path::{Path, PathBuf},
     sync::atomic::AtomicBool,
     time::{Duration, Instant},
 };
@@ -175,4 +176,18 @@ where
         .name(format!("Seelen Thread - {}", id))
         .spawn(cb)
         .map_err(|e| format!("Failed to spawn thread: {}", e).into())
+}
+
+pub fn copy_dir_all(src: impl AsRef<Path>, dst: impl AsRef<Path>) -> Result<()> {
+    fs::create_dir_all(&dst)?;
+    for entry in fs::read_dir(src)? {
+        let entry = entry?;
+        let ty = entry.file_type()?;
+        if ty.is_dir() {
+            copy_dir_all(entry.path(), dst.as_ref().join(entry.file_name()))?;
+        } else {
+            fs::copy(entry.path(), dst.as_ref().join(entry.file_name()))?;
+        }
+    }
+    Ok(())
 }